requirements.txt
Django~=2.2.4
mysqlclient==1.4.6
djangorestframework==3.11.0
django-cors-headers==3.5.0
django-rest-swagger==2.2.0
pytz==2020.1
sqlparse==0.3.1
Pillow==7.2.0
drf-yasg==1.17.1
Pyrebase==3.0.27